Refactored search bar
