+++ src/components/home/searchBar/SearchInfo.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const SearchInfo: React.FunctionComponent = () => {
+	return (
+		<>
+			<h3>Search through over 2.3 million recipes from all over the web!</h3>
+			<p>Click on a recipe card to see the ingredients and calories!</p>
+			<p>Click the link on the recipe card to visit the recipe page!</p>
+		</>
+	);
+};
+
+export { SearchInfo };
